blogdeploy: Override url instead of baseurl when deploying unstable
Blogdeploy overrides the baseurl of the jekyll configuration when
deploying from unstable (which is deployed to /unstable), which
causes page links to be correct, but other URLs such as those for
images to be incorrect.

This commit modifies blogdeploy to append "/unstable" to the url
instead of overriding the baseurl.